Convert colours to use Word8 rather than Double
This should make the library better in the long run, especially for
image analysis (i.e. histograms), but there are some issues with it
currently.

HSV and HSL will be left using Double for now for simplicity's sake.

Convolution no longer works for most colour types, that will need to be
fixed as soon as possible.

Gamma correction is no longer exactly possible (as seen in `Main.hs`) so
that will also need to be addressed soon (alongside other similar point
processes).

Generating images from functions is also a bit more complicated now so I
will try to improve that somehow.