Fix inconsistent cropping when centre not possible
Given an odd size to crop to and an even-sized image, `cropToSize` took
an extra row from below (row+1) the centre and an extra column from the
right (col+1).
Given an even size to crop to and an odd-sized image, it took a row from
above (row-1) and a column from the left (col-1).

This is inconsistent and could easily lead to confusion.
Therefore, with this commit, `cropToSize` now always takes from above
(row-1) and from the left (col-1), as needed.

This commit also changes `zoomToSize` to work in the same way, in order
to maximise consistency throughout the library.