Implement functions for generating multi-channel histograms
All use the same (mostly unchanged) naive and sequential method, so they
are very inefficient and slow.
This will be addressed in the future.

An example usage of the functions is in `app/main.hs`, where it will
output a visual histogram (in the form of a bar chart) showing the
amount of values at each RGB and grayscale value.